requirements.txt
google-auth>=2.47.0
gspread>=6.2.1
pandas>=2.3.3
plotly>=6.5.2
pydantic>=2.12.5
python-dotenv>=1.0.0
streamlit>=1.53.0
